\name{twMiscRgl-package}
\alias{twMiscRgl-package}
\alias{twMiscRgl}
\docType{package}
\title{tw 3D utility functions}
\description{utilities used across several other packages of tw. Uses rgl and misc3d packages and therefore not in twMisc}
\details{
\tabular{ll}{Package: \tab twMiscRgl\cr
Title: \tab tw 3D utility functions\cr
Version: \tab 0.5-1\cr
Date: \tab 2010-11-26\cr
Author: \tab Thomas Wutzler\cr
Maintainer: \tab Thomas Wutzler <twutz@bgc-jena.mpg.de>\cr
Depends: \tab twMisc, rgl\cr
Imports: \tab misc3d\cr
Suggests: \tab inlinedocs,debug\cr
License: \tab Apache License (== 2.0)\cr
LazyLoad: \tab yes\cr}
}
\author{Thomas Wutzler}

\keyword{ package }


